gfx: Split into multiple headers
