+++ gfx/gfx_example.cpp
@@ -2,7 +2,8 @@
 //
 // SPDX-License-Identifier: BSD-2-Clause
 
-#include "gfx/gfx.h"
+#include "gfx/color.h"
+#include "gfx/opengl_painter.h"
 
 #include <SFML/Window.hpp>
