+++ gfx/opengl_painter.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: BSD-2-Clause
 
-#include "gfx/gfx.h"
+#include "gfx/opengl_painter.h"
 
 #include <GL/glew.h>
